Added download handlers.
